+++ flutter_service_bloc/example/lib/main.dart
@@ -124,7 +124,7 @@ class OpenLibraryAuthorSearchServiceBloc extends PaginationServiceBloc<
           pagination: NumberBasedPagination(
             onUpdateHasNextPage: (responseData) => responseData.isNotEmpty,
           ),
-          dataProcessor: PaginationListResponseDataProcessor(),
+          paginationResponseData: PaginationListResponseData(),
         );
 
   final OpenLibraryRepository repository;